pyproject.toml
[project]
name = "supabase-python-pipeline"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openpyxl>=3.1.5",
    "pandas>=2.3.2",
    "plotly>=6.3.0",
    "python-dotenv>=1.1.1",
    "rich>=14.1.0",
    "seaborn>=0.13.2",
    "streamlit>=1.49.1",
    "supabase>=2.18.1",
    "xlsxwriter>=3.2.5",
]
